config/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
python-dotenv==1.2.1
psycopg2-binary==2.9.11
pgvector==0.4.2

langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
openai>=1.54.0,<2.0.0

PyMuPDF==1.27.1
python-docx==1.2.0
lxml==6.0.2
